TabooButtonCore 적용
